refactor: extract constants in separate file
