+++ lib/pbxFile.js
@@ -16,134 +16,24 @@
  */
 
 var path = require('path'),
-    util = require('util');
-
-var DEFAULT_SOURCETREE = '"<group>"',
-    DEFAULT_PRODUCT_SOURCETREE = 'BUILT_PRODUCTS_DIR',
-    DEFAULT_FILEENCODING = 4,
-    DEFAULT_GROUP = 'Resources',
-    DEFAULT_FILETYPE = 'unknown',
-    HEADER_FILE_TYPE_SUFFIX = ".h",
-    ENTITLEMENTS_FILE_TYPE_SUFFIX = ".entitlements",
-    SOURCE_CODE_FILE_TYPE_PREFIX = "sourcecode.";
-
-var FILETYPE_BY_EXTENSION = {
-        a: 'archive.ar',
-        app: 'wrapper.application',
-        appex: 'wrapper.app-extension',
-        bundle: 'wrapper.plug-in',
-        c: 'sourcecode.c.c',
-        cc: 'sourcecode.cpp.cpp',
-        cpp: 'sourcecode.cpp.cpp',
-        cxx: 'sourcecode.cpp.cpp',
-        'c++': 'sourcecode.cpp.cpp',
-        dylib: 'compiled.mach-o.dylib',
-        framework: 'wrapper.framework',
-        h: 'sourcecode.c.h',
-        hh: 'sourcecode.cpp.h',
-        hpp: 'sourcecode.cpp.h',
-        hxx: 'sourcecode.cpp.h',
-        'h++': 'sourcecode.cpp.h',
-        m: 'sourcecode.c.objc',
-        mm: 'sourcecode.cpp.objcpp',
-        markdown: 'text',
-        mdimporter: 'wrapper.cfbundle',
-        octest: 'wrapper.cfbundle',
-        pch: 'sourcecode.c.h',
-        plist: 'text.plist.xml',
-        entitlements: 'text.plist.entitlements',
-        png: "image.png",
-        sh: 'text.script.sh',
-        swift: 'sourcecode.swift',
-        tbd: 'sourcecode.text-based-dylib-definition',
-        xcassets: 'folder.assetcatalog',
-        xcconfig: 'text.xcconfig',
-        xcdatamodel: 'wrapper.xcdatamodel',
-        xcodeproj: 'wrapper.pb-project',
-        xctest: 'wrapper.cfbundle',
-        xib: 'file.xib',
-        strings: 'text.plist.strings'
-    },
-    GROUP_BY_FILETYPE = {
-        'archive.ar': 'Frameworks',
-        'compiled.mach-o.dylib': 'Frameworks',
-        'sourcecode.text-based-dylib-definition': 'Frameworks',
-        'wrapper.framework': 'Frameworks',
-        'embedded.framework': 'Embed Frameworks',
-        'sourcecode.c.h': 'Resources',
-        'sourcecode.c.c': 'Sources',
-        'sourcecode.c.objc': 'Sources',
-        'sourcecode.swift': 'Sources',
-        'sourcecode.cpp.cpp': 'Sources',
-        'sourcecode.cpp.objcpp': 'Sources'
-    },
-    PATH_BY_FILETYPE = {
-        'compiled.mach-o.dylib': 'usr/lib/',
-        'sourcecode.text-based-dylib-definition': 'usr/lib/',
-        'wrapper.framework': 'System/Library/Frameworks/'
-    },
-    SOURCETREE_BY_FILETYPE = {
-        'compiled.mach-o.dylib': 'SDKROOT',
-        'sourcecode.text-based-dylib-definition': 'SDKROOT',
-        'wrapper.framework': 'SDKROOT'
-    },
-    ENCODING_BY_FILETYPE = {
-        'sourcecode.c.h': 4,
-        'sourcecode.c.h': 4,
-        'sourcecode.cpp.h': 4,
-        'sourcecode.c.c': 4,
-        'sourcecode.c.objc': 4,
-        'sourcecode.cpp.cpp': 4,
-        'sourcecode.cpp.objcpp': 4,
-        'sourcecode.swift': 4,
-        'text': 4,
-        'text.plist.xml': 4,
-        'text.script.sh': 4,
-        'text.xcconfig': 4,
-        'text.plist.strings': 4
-    };
-
-function isHeaderFileType(fileType) {
-    return fileType.endsWith(HEADER_FILE_TYPE_SUFFIX);
-}
-
-function isSourceFileType(fileType) {
-    return fileType.startsWith(SOURCE_CODE_FILE_TYPE_PREFIX) && !isHeaderFileType(fileType);
-}
-
-function isAssetFileType(fileType) {
-    return fileType === FILETYPE_BY_EXTENSION.xcassets;
-}
-
-function isResource(group) {
-    return group === "Resources";
-}
-
-function isEntitlement(fileType) {
-    return fileType.endsWith(ENTITLEMENTS_FILE_TYPE_SUFFIX);
-}
-
-function isPlist(fileType) {
-    return fileType === FILETYPE_BY_EXTENSION.plist;
-}
-
-function unquoted(text){
-    return text == null ? '' : text.replace (/(^")|("$)/g, '')
-}
+    util = require('util'),
+    constants = require('./constants'),
+    unquoted = constants.unquoted,
+    FILETYPE_BY_EXTENSION = constants.FILETYPE_BY_EXTENSION;
 
 function detectType(filePath) {
     var extension = path.extname(filePath).substring(1),
         filetype = FILETYPE_BY_EXTENSION[unquoted(extension)];
 
     if (!filetype) {
-        return DEFAULT_FILETYPE;
+        return constants.DEFAULT_FILETYPE;
     }
 
     return filetype;
 }
 
 function defaultExtension(fileRef) {
-    var filetype = fileRef.lastKnownFileType && fileRef.lastKnownFileType != DEFAULT_FILETYPE ?
+    var filetype = fileRef.lastKnownFileType && fileRef.lastKnownFileType != constants.DEFAULT_FILETYPE ?
         fileRef.lastKnownFileType : fileRef.explicitFileType;
 
     for(var extension in FILETYPE_BY_EXTENSION) {
@@ -156,7 +46,7 @@ function defaultExtension(fileRef) {
 
 function defaultEncoding(fileRef) {
     var filetype = fileRef.lastKnownFileType || fileRef.explicitFileType,
-        encoding = ENCODING_BY_FILETYPE[unquoted(filetype)];
+        encoding = constants.ENCODING_BY_FILETYPE[unquoted(filetype)];
 
     if (encoding) {
         return encoding;
@@ -166,18 +56,18 @@ function defaultEncoding(fileRef) {
 function detectGroup(fileRef, opt) {
     var extension = path.extname(fileRef.basename).substring(1),
         filetype = fileRef.lastKnownFileType || fileRef.explicitFileType,
-        groupName = GROUP_BY_FILETYPE[unquoted(filetype)];
+        groupName = constants.GROUP_BY_FILETYPE[unquoted(filetype)];
 
     if (extension === 'xcdatamodeld') {
         return 'Sources';
     }
 
     if (opt.customFramework && opt.embed) {
-        return GROUP_BY_FILETYPE['embedded.framework'];
+        return constants.GROUP_BY_FILETYPE['embedded.framework'];
     }
 
     if (!groupName) {
-        return DEFAULT_GROUP;
+        return constants.DEFAULT_GROUP;
     }
 
     return groupName;
@@ -186,18 +76,18 @@ function detectGroup(fileRef, opt) {
 function detectSourcetree(fileRef) {
 
     var filetype = fileRef.lastKnownFileType || fileRef.explicitFileType,
-        sourcetree = SOURCETREE_BY_FILETYPE[unquoted(filetype)];
+        sourcetree = constants.SOURCETREE_BY_FILETYPE[unquoted(filetype)];
 
     if (fileRef.explicitFileType) {
-        return DEFAULT_PRODUCT_SOURCETREE;
+        return constants.DEFAULT_PRODUCT_SOURCETREE;
     }
 
     if (fileRef.customFramework) {
-        return DEFAULT_SOURCETREE;
+        return constants.DEFAULT_SOURCETREE;
     }
 
     if (!sourcetree) {
-        return DEFAULT_SOURCETREE;
+        return constants.DEFAULT_SOURCETREE;
     }
 
     return sourcetree;
@@ -205,7 +95,7 @@ function detectSourcetree(fileRef) {
 
 function defaultPath(fileRef, filePath) {
     var filetype = fileRef.lastKnownFileType || fileRef.explicitFileType,
-        defaultPath = PATH_BY_FILETYPE[unquoted(filetype)];
+        defaultPath = constants.PATH_BY_FILETYPE[unquoted(filetype)];
 
     if (fileRef.customFramework) {
         return filePath;
@@ -218,18 +108,8 @@ function defaultPath(fileRef, filePath) {
     return filePath;
 }
 
-function defaultGroup(fileRef) {
-    var groupName = GROUP_BY_FILETYPE[fileRef.lastKnownFileType];
-
-    if (!groupName) {
-        return DEFAULT_GROUP;
-    }
-
-    return defaultGroup;
-}
-
 function pbxFile(filepath, opt) {
-    var opt = opt || {};
+    opt = opt || {};
 
     this.basename = opt.basename || path.basename(filepath);
     this.lastKnownFileType = opt.lastKnownFileType || detectType(filepath);
@@ -275,12 +155,4 @@ function pbxFile(filepath, opt) {
     }
 }
 
-module.exports = {
-    pbxFile: pbxFile,
-    isSourceFileType,
-    isHeaderFileType,
-    isResource,
-    isEntitlement,
-    isAssetFileType,
-    isPlist
-}
+module.exports = pbxFile;
