+++ lib/pbxProject.js
@@ -22,18 +22,19 @@ var util = require('util'),
     $uuid = require('uuid'),
     fork = require('child_process').fork,
     pbxWriter = require('./pbxWriter'),
-    pbxFile = require('./pbxFile').pbxFile,
-    isSourceFileType = require('./pbxFile').isSourceFileType,
-    isHeaderFileType = require('./pbxFile').isHeaderFileType,
-    isResource = require('./pbxFile').isResource,
-    isEntitlement = require('./pbxFile').isEntitlement,
-    isAssetFileType = require('./pbxFile').isAssetFileType,
-    isPlist = require('./pbxFile').isPlist,
+    pbxFile = require('./pbxFile'),
+    constants = require('./constants'),
     fs = require('fs'),
     parser = require('./parser/pbxproj'),
     plist = require('simple-plist'),
     COMMENT_KEY = /_comment$/,
-    NO_SPECIAL_SYMBOLS = /^[a-zA-Z0-9_\.\$]+\.[a-zA-Z]+$/;
+    NO_SPECIAL_SYMBOLS = /^[a-zA-Z0-9_\.\$]+\.[a-zA-Z]+$/,
+    isSourceFileType = constants.isSourceFileType,
+    isHeaderFileType = constants.isHeaderFileType,
+    isResource = constants.isResource,
+    isEntitlementFileType = constants.isEntitlementFileType,
+    isAssetFileType = constants.isAssetFileType,
+    isPlistFileType = constants.isPlistFileType;
 
 function pbxProject(filename) {
     if (!(this instanceof pbxProject))
@@ -81,7 +82,7 @@ pbxProject.prototype.allUuids = function() {
         uuids = [],
         section;
 
-    for (key in sections) {
+    for (var key in sections) {
         section = sections[key]
         uuids = uuids.concat(Object.keys(section))
     }
@@ -440,7 +441,7 @@ pbxProject.prototype.removeCopyfile = function(fpath, opt) {
 
 pbxProject.prototype.removeFromPbxCopyfilesBuildPhase = function(file) {
     var sources = this.pbxCopyfilesBuildPhaseObj(file.target);
-    for (i in sources.files) {
+    for (var i in sources.files) {
         if (sources.files[i].comment == longComment(file)) {
             sources.files.splice(i, 1);
             break;
@@ -581,11 +582,11 @@ pbxProject.prototype.addPbxGroup = function (filePathsArray, name, path, sourceT
         } else {
             this.addToPbxFileReferenceSection(file);    // PBXFileReference
             pbxGroup.children.push(pbxGroupChild(file));
-            if(isHeaderFileType(file.lastKnownFileType) || isPlist(file.lastKnownFileType)) {
+            if(isHeaderFileType(file.lastKnownFileType) || isPlistFileType(file.lastKnownFileType)) {
                 continue;
             }
 
-            if(isEntitlement(file.lastKnownFileType)) {
+            if(isEntitlementFileType(file.lastKnownFileType)) {
                 this.addToBuildSettings('CODE_SIGN_ENTITLEMENTS', file.path, opt.target);
                 continue;
             }
@@ -839,7 +840,7 @@ pbxProject.prototype.removeFromFrameworksPbxGroup = function(file) {
     }
     var pluginsGroupChildren = this.pbxGroupByName('Frameworks').children;
 
-    for (i in pluginsGroupChildren) {
+    for (var i in pluginsGroupChildren) {
         if (pbxGroupChild(file).value == pluginsGroupChildren[i].value &&
             pbxGroupChild(file).comment == pluginsGroupChildren[i].comment) {
             pluginsGroupChildren.splice(i, 1);
@@ -859,7 +860,7 @@ pbxProject.prototype.removeFromPbxEmbedFrameworksBuildPhase = function (file) {
     var sources = this.pbxEmbedFrameworksBuildPhaseObj(file.target);
     if (sources) {
         var files = [];
-        for (i in sources.files) {
+        for (var i in sources.files) {
             if (sources.files[i].comment != longComment(file)) {
                 files.push(sources.files[i]);
             }
@@ -930,7 +931,7 @@ pbxProject.prototype.addToPbxFrameworksBuildPhase = function(file) {
 
 pbxProject.prototype.removeFromPbxFrameworksBuildPhase = function(file) {
     var sources = this.pbxFrameworksBuildPhaseObj(file.target);
-    for (i in sources.files) {
+    for (var i in sources.files) {
         if (sources.files[i].comment == longComment(file)) {
             sources.files.splice(i, 1);
             break;
@@ -1075,8 +1076,8 @@ pbxProject.prototype.addBuildPhase = function(filePathsArray, buildPhaseType, co
         if (!COMMENT_KEY.test(key)) continue;
 
         var buildFileKey = key.split(COMMENT_KEY)[0],
-            buildFile = buildFileSection[buildFileKey];
-        fileReference = fileReferenceSection[buildFile.fileRef];
+            buildFile = buildFileSection[buildFileKey],
+            fileReference = fileReferenceSection[buildFile.fileRef];
 
         if (!fileReference) continue;
 
@@ -1649,7 +1650,7 @@ pbxProject.prototype.addTarget = function(name, type, subfolder) {
 
        // this.addBuildPhaseToTarget(newPhase.buildPhase, this.getFirstTarget().uuid)
 
-    };
+    }
 
     // Target: Add uuid to root project
     this.addToPbxProjectSection(target);
@@ -1684,13 +1685,13 @@ pbxProject.prototype.removeTarget = function(target, targetKey) {
         var sectionUuid = buildPhase.value;
         var section = {}; //in case we don't recognise the section
         if(buildPhase.comment === buildPhaseNameForIsa("PBXSourcesBuildPhase")){
-            var section = this.hash.project.objects["PBXSourcesBuildPhase"];
+            section = this.hash.project.objects["PBXSourcesBuildPhase"];
             files = files.concat(section[sectionUuid].files);
         } else if (buildPhase.comment === buildPhaseNameForIsa("PBXResourcesBuildPhase")) {
-            var section = this.hash.project.objects["PBXResourcesBuildPhase"];
+            section = this.hash.project.objects["PBXResourcesBuildPhase"];
             files = files.concat(section[sectionUuid].files);
         } else if (buildPhase.comment === buildPhaseNameForIsa("PBXFrameworksBuildPhase")) {
-            var section = this.hash.project.objects["PBXFrameworksBuildPhase"];
+            section = this.hash.project.objects["PBXFrameworksBuildPhase"];
             files = files.concat(section[sectionUuid].files);
         }
 
@@ -1737,7 +1738,7 @@ pbxProject.prototype.removeTarget = function(target, targetKey) {
 
     // the productReference is the uuid from the PBXFileReference Section, but we need the one in PBXBuildFile section
     // check the fileRef of all records until we find the product
-    for (uuid in pbxBuildFileSectionNoComments) {
+    for (var uuid in pbxBuildFileSectionNoComments) {
         if (this.pbxBuildFileSection()[uuid].fileRef == productReferenceUuid) {
             productUuid = uuid;
         }
@@ -1804,13 +1805,13 @@ pbxProject.prototype.removeTarget = function(target, targetKey) {
     var pbxContainerItemProxySection = this.hash.project.objects[pbxContainerItemProxy];
 
     for(var targetDependencyUuid in pbxTargetDependencySectionNoComments) {
-        targetDependency = pbxTargetDependencySectionNoComments[targetDependencyUuid];
+        var targetDependency = pbxTargetDependencySectionNoComments[targetDependencyUuid];
         if(targetDependency.target === targetKey) {
             //remove the PBXContainerItemProxy
             removeItemAndCommentFromSectionByUuid(pbxContainerItemProxySection, targetDependency.targetProxy);
             //remove the PBXTargetDependency from dependencies from all targets
-            for (var nativeTargetUuid in nativeTargetsnoComments) {
-                const nativeTarget = nativeTargetsnoComments[nativeTargetUuid];
+            for (var nativeTargetKey in nativeTargetsnoComments) {
+                const nativeTarget = nativeTargetsnoComments[nativeTargetKey];
                 for (var dependencyIndex in nativeTarget.dependencies) {
                     if (nativeTarget.dependencies[dependencyIndex].value == targetDependencyUuid) {
                         nativeTarget.dependencies.splice(dependencyIndex, 1);
@@ -2061,7 +2062,7 @@ function unquote(str) {
 
 function buildPhaseNameForIsa (isa) {
 
-    BUILDPHASENAME_BY_ISA = {
+    var BUILDPHASENAME_BY_ISA = {
         PBXCopyFilesBuildPhase: 'Copy Files',
         PBXResourcesBuildPhase: 'Resources',
         PBXSourcesBuildPhase: 'Sources',
@@ -2073,7 +2074,7 @@ function buildPhaseNameForIsa (isa) {
 
 function producttypeForTargettype (targetType) {
 
-    PRODUCTTYPE_BY_TARGETTYPE = {
+    var PRODUCTTYPE_BY_TARGETTYPE = {
             application: 'com.apple.product-type.application',
             app_extension: 'com.apple.product-type.app-extension',
             bundle: 'com.apple.product-type.bundle',
@@ -2091,7 +2092,7 @@ function producttypeForTargettype (targetType) {
 
 function filetypeForProducttype (productType) {
 
-    FILETYPE_BY_PRODUCTTYPE = {
+    var FILETYPE_BY_PRODUCTTYPE = {
             'com.apple.product-type.application': '"wrapper.application"',
             'com.apple.product-type.app-extension': '"wrapper.app-extension"',
             'com.apple.product-type.bundle': '"wrapper.plug-in"',
