+++ test/addFramework.js
@@ -18,7 +18,7 @@
 var fullProject = require('./fixtures/full-project')
     fullProjectStr = JSON.stringify(fullProject),
     pbx = require('../lib/pbxProject'),
-    pbxFile = require('../lib/pbxFile').pbxFile,
+    pbxFile = require('../lib/pbxFile'),
     proj = new pbx('.');
 
 function cleanHash() {
