+++ test/dataModelDocument.js
@@ -19,7 +19,7 @@ var jsonProject = require('./fixtures/full-project')
     fullProjectStr = JSON.stringify(jsonProject),
     path = require('path'),
     pbx = require('../lib/pbxProject'),
-    pbxFile = require('../lib/pbxFile').pbxFile,
+    pbxFile = require('../lib/pbxFile'),
     proj = new pbx('.'),
     singleDataModelFilePath = __dirname + '/fixtures/single-data-model.xcdatamodeld',
     multipleDataModelFilePath = __dirname + '/fixtures/multiple-data-model.xcdatamodeld';
