+++ test/pbxFile.js
@@ -15,7 +15,7 @@
  under the License.
  */
 
-var pbxFile = require('../lib/pbxFile').pbxFile;
+var pbxFile = require('../lib/pbxFile');
 
 exports['lastKnownFileType'] = {
     'should detect that a .m path means sourcecode.c.objc': function (test) {
