+++ test/xcode5searchPaths.js
@@ -18,7 +18,7 @@
 var xcode5proj = require('./fixtures/library-search-paths')
     xcode5projStr = JSON.stringify(xcode5proj),
     pbx = require('../lib/pbxProject'),
-    pbxFile = require('../lib/pbxFile').pbxFile,
+    pbxFile = require('../lib/pbxFile'),
     proj = new pbx('.'),
     libPoop = { path: 'some/path/poop.a' };
